requirements.txt
python-dotenv
pandas 
numpy
scikit-learn
certifi
mlflow<3
pymongo
dill
pyaml
dagshub
fastapi
uvicorn
python-multipart
##-e .